added file checking to install script

+++ scripts/install.sh
@@ -1,18 +1,34 @@
 !/bin/bash
 
 cd ~
-git clone https://github.com/TimoLob/dotfiles.git --recurse-submodules
+git clone https://github.com/TimoLob/dotfiles.git 
 
-yay -S kitty zsh stow ttf-meslo-nerd-font-powerlevel10k zsh-theme-powerlevel10k-git ripgrep zoxide neovim-git --noconfirm
+yay -S kitty zsh stow ttf-meslo-nerd-font-powerlevel10k zsh-theme-powerlevel10k-git ripgrep fd zoxide neovim-git --noconfirm
 
 
-mv ~/.config/kitty ~/.config/kitty_bak
-mv ~/.config/i3 ~/.config/i3_bak
-mv ~/.config/nvim ~/.config/nvim_bak
-mv ~/.zshrc ~/.zshrc_bak
-mv ~/.backgrounds ~/.backgrounds_bak
+if [ -d "~/.config/kitty" ]; then
+	mv ~/.config/kitty ~/.config/kitty_bak
+fi
+
+if [ -d "~/.config/i3" ]; then
+	mv ~/.config/i3 ~/.config/i3_bak
+fi
+
+if [ -d "~/.config/nvim" ]; then
+	mv ~/.config/nvim ~/.config/nvim_bak
+fi
+
+if [ -d "~/.backgrounds" ]; then
+	mv ~/.backgrounds ~/.backgrounds_bak
+fi
+
+if [ -f "~/.zhsrc" ]; then
+	mv ~/.zshrc ~/.zshrc_bak
+fi
 
 cd dotfiles
+git clone https://github.com/TimoLob/my_neovim_config.git .config/nvim
+
 stow .
 cd scripts
 source ./set_bg.sh
